Remove automatic partial correlation calculation
Partial correlation matrices for atlas-based connectivity can only be
calculated if the covariance of all region time series is invertible.
This is usually the case if there are more time points than regions. In
an earlier version, when this what not the case, the user would get an
numpy.linalg.linalg.LinAlgError, which is quite confusing. As a stop
gap, I added a try-catch clause to convert the error into a warning.
However, after careful discussion, we decided to remove partial
correlation calculation completely, as this is a trivial step to do in
postprocessing.