Fix error when using `--aggregate` flag in `group-level`
For the IMAGEN dataset, subjects were scanned in to sessions `A` and
`B`, but each session only has one scan for each task. That means that
we want to use `--aggregate ses` to merge the sessions before running
group-level statistics, but there will be no actual fixed effects
aggregations to run. This commit fixes the error that was raised in
this case.

Reported-by: Susanne Erk <susanne.erk@charite.de>