Fix URLs and start migration to conda for build (#702)
This pull request addresses issue [#697](https://github.com/HALFpipe/HALFpipe/issues/697 ). All of the python dependencies have been either found in conga-forge in another number for the version, or with another name. The only python package exception is `sqlalchemy` where for some reason pip continues to install it even though there is a recipe available in condaforge. 

The reason we want to migrate entirely to conda management is to simplify the `requirements.sh` file and the build of the `Dockerfile`. The next pull request will address the remaining packages, for which we expect to build a custom recipe for all.

* Fix PNAS_Smith09_rsn10.nii.gz download link

* Conda recipe for niflow, improved dockerfile phrasing, deprecated line_profiler

* Replace buildkit syntax with a copy instruction

* Typo was breaking build for niflow

* Control all pypi based python dependencies through conda

* Tzdata was breaking build because it  corresponds to a different package in conda

* Upgrade pybids to the first version submitted in conda

* Roll back bids as upgrade breaks tests, include nitransforms recipe

* recipe pybids