chore(ci): build deps outside of dockerfile and add gica_cmd recipes (#815)
Extend the environment in the HALFpipe container to include the MATLAB compiler runtime and packages compiled by the MATLAB compiler. This is a prerequisite for adding NeuroMark and other MATLAB-based analyses down the line. As MATLAB is a complex dependency, we move the dependency builds to not be part of the Dockerfile anymore.

* chore(ci): build deps script added

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* chore(ci): move secrets to correct step

* Add recipes

* Fix yaml syntax

* Set gica_cmd version

* Build packages outside of the Dockerfile

* Fix micromamba setup

* Initialize micromamba environment in workflow steps

as per https://github.com/mamba-org/setup-micromamba?tab=readme-ov-file#about-login-shells

* Only upload built packages on push

* Fix version patching logic

* Fix mypy errors

* Automatically cancel previous runs on push

* Remove unneeded secrets

* Fix mypy error

* Make matlab license available to runner

* Add gica_cmd dependency and pin pcre version

* Fix building afni

* Fix pinning via variant files

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Lea Waller <lea.waller@charite.de>